Add api to run windows CMD command lines via ssh
The aggregation of the commands is done by concatenate the commands using && operands
and wrapping it by (), so each commads which fails will stop the execution flow.
Linux script API is not changed due to backwards compatibility.

Tested by executing on Windows8.1, both single and mutiple commands, with expected
failures and without.